pyproject.toml
[tool.poetry]
name = "onlinestore"
version = "0.0.1"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.1"
django = "^4.2.3"
beautifulsoup4 = "^4.12.2"
celery = "^5.3.1"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
django-filter = "^23.2"
django-rest-passwordreset = "^1.3.0"
psycopg2-binary = "^2.9.6"
drf-yasg = "^1.21.7"
django-uuslug = "^2.0.0"
django-mptt = "^0.14.0"
whitenoise = "^6.5.0"
coreapi = "^2.3.3"
djangorestframework-simplejwt = "^5.2.2"
gunicorn = "^21.2.0"
braintree = "^4.21.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pre-commit = "^3.3.3"
flake8-bugbear = "^23.7.10"
flake8-simplify = "^0.20.0"
flake8-builtins = "^2.1.0"
pep8-naming = "^0.13.3"
flake8-import-order = "^0.18.2"

[tool.black]
line-length = 120
skip-string-normalization = true

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

# doesn't work with pre-commit
[tool.docformatter]
wrap-summaries = 120
wrap-descriptions = 120

[pycodestyle]
max-line-length = 120
exclude = ".tox,.git,*/migrations/*,*/static/CACHE/*,docs,node_modules"
